Updated line recording
Now the reset button resets the scan size to the original and no longer removes the annotations made. The delete function works directily witht he saved current lines and removes only what it is told. a new function called remove_sequence was added in aiding of removing the annotations chosen by the user